+++ setup.php
@@ -27,13 +27,19 @@
  * @link      https://github.com/pluginsGLPI/tag
  * -------------------------------------------------------------------------
  */
+
+use Glpi\Application\ImportMapGenerator;
+use Glpi\Form\Destination\AbstractCommonITILFormDestination;
+use Glpi\Form\Destination\FormDestinationManager;
 use Glpi\Form\Form;
+use Glpi\Form\Migration\TypesConversionMapper;
+use Glpi\Form\QuestionType\QuestionTypesManager;
 use Glpi\Plugin\Hooks;
 
 define('PLUGIN_TAG_VERSION', '2.13.0');
 
 // Minimal GLPI version, inclusive
-define("PLUGIN_TAG_MIN_GLPI", "11.0.0");
+define("PLUGIN_TAG_MIN_GLPI", "11.0.1");
 // Maximum GLPI version, exclusive
 define("PLUGIN_TAG_MAX_GLPI", "11.0.99");
 
@@ -150,10 +156,15 @@ function plugin_init_tag()
             $PLUGIN_HOOKS[Hooks::ADD_JAVASCRIPT]['tag'][] = 'js/entity.js';
         }
 
+        ImportMapGenerator::getInstance()->registerModulesPath('tag', '/public/js/modules');
+
         Plugin::registerClass('PluginTagProfile', ['addtabon' => ['Profile']]);
         Plugin::registerClass('PluginTagConfig', ['addtabon' => 'Config']);
 
         $PLUGIN_HOOKS['use_rules']['tag']      = ['RuleTicket'];
+
+        // Register tag question type
+        registerPluginTypes();
     }
 }
 
@@ -202,4 +213,26 @@ function plugin_tag_geturl(): string
     /** @var array $CFG_GLPI */
     global $CFG_GLPI;
     return sprintf('%s/plugins/tag', $CFG_GLPI['url_base']);
+}
+
+function registerPluginTypes(): void
+{
+    $types = QuestionTypesManager::getInstance();
+    $type_mapper = TypesConversionMapper::getInstance();
+    $destination_manager = FormDestinationManager::getInstance();
+
+    // Register question type category
+    $types->registerPluginCategory(new PluginTagQuestionTypeCategory());
+
+    // Register question type
+    $types->registerPluginQuestionType(new PluginTagQuestionType());
+
+    // Register question type itil fields
+    $destination_manager->registerPluginCommonITILConfigField(
+        AbstractCommonITILFormDestination::class,
+        new PluginTagDestinationField(),
+    );
+
+    // Register mapper for legacy question type
+    $type_mapper->registerPluginQuestionTypeConverter('tag', new PluginTagQuestionType());
 }